client side refactoring ok
