+++ client/src/components/people/withPeopleProps.js
@@ -1,7 +1,7 @@
 import { withProps } from "libraries/model";
 import { usersSelector } from "models/users/props";
 import { fetchUsers } from "models/users/actions";
-import { deleteUser } from "models/user/actions";
+import { deleteUser } from "models/users/actions";
 
 const withPeopleProps = withProps(
   { usersSelector },
