+++ client/src/components/people/withPeoplePropsHooks.js
@@ -2,7 +2,7 @@ import React, { memo } from "react";
 import { useProps } from "libraries/model";
 import { usersSelector } from "models/users/props";
 import { fetchUsers } from "models/users/actions";
-import { deleteUser } from "models/user/actions";
+import { deleteUser } from "models/users/actions";
 
 const withPeopleProps = Component => {
   const MemoizedComponent = memo(Component);
