+++ client/src/models/user/index.js
@@ -1,5 +1,5 @@
 import * as actions from "./actions";
-import * as epics from "./epics";
+import epics from "./epics";
 import { userReducer as reducer } from "./reducers";
 
 export { actions, epics, reducer };
