+++ client/src/models/user/reducers.js
@@ -1,12 +1,5 @@
-import {
-  fetchUser,
-  fetchUserFulfilled,
-  fetchUserRejected,
-  deleteUser,
-  deleteUserFulfilled,
-  deleteUserRejected
-} from "./actions";
-import { user, afterUserDeleted } from "./props";
+import { fetchUser, fetchUserFulfilled, fetchUserRejected } from "./actions";
+import { user } from "./props";
 
 export const userReducer = (state = {}, action) => {
   switch (action.type) {
@@ -15,12 +8,6 @@ export const userReducer = (state = {}, action) => {
     case fetchUserFulfilled.type:
       return user(state, action.payload);
     case fetchUserRejected.type:
-      return action.payload;
-    case deleteUser.type:
-      return state; //need state for the omit function in DELETE_USER_FULFILLED
-    case deleteUserFulfilled.type:
-      return afterUserDeleted(state, action.payload);
-    case deleteUserRejected.type:
       return action.payload;
     default:
       return state;
