+++ client/src/models/users/actions.js
@@ -6,4 +6,15 @@ const fetchUsers = actionCreator(`FETCH_${CMP}`);
 const fetchUsersFulfilled = actionCreator(`FETCH_${CMP}_FULFILLED`);
 const fetchUsersRejected = actionCreator(`FETCH_${CMP}_REJECTED`);
 
-export { fetchUsers, fetchUsersFulfilled, fetchUsersRejected };
+const deleteUser = actionCreator("DELETE_USER");
+const deleteUserFulfilled = actionCreator("DELETE_USER_FULFILLED");
+const deleteUserRejected = actionCreator("DELETE_USER_REJECTED");
+
+export {
+  fetchUsers,
+  fetchUsersFulfilled,
+  fetchUsersRejected,
+  deleteUser,
+  deleteUserFulfilled,
+  deleteUserRejected
+};
