+++ client/src/models/users/epics.js
@@ -2,9 +2,16 @@ import { of } from "rxjs";
 import { ajax } from "rxjs/ajax";
 import { map, catchError, mergeMap } from "rxjs/operators";
 import { ofType } from "redux-observable";
-import { fetchUsers, fetchUsersFulfilled, fetchUsersRejected } from "./actions";
+import {
+  fetchUsers,
+  fetchUsersFulfilled,
+  fetchUsersRejected,
+  deleteUser,
+  deleteUserFulfilled,
+  deleteUserRejected
+} from "./actions";
 
-export default action$ =>
+const fetchUsersEpic = action$ =>
   action$.pipe(
     ofType(fetchUsers.type),
     mergeMap(() =>
@@ -19,3 +26,21 @@ export default action$ =>
       )
     )
   );
+
+const deleteUserEpic = action$ =>
+  action$.pipe(
+    ofType(deleteUser.type),
+    mergeMap(action =>
+      ajax.delete(`http://localhost:3000/api/users/${action.payload}`).pipe(
+        map(response => deleteUserFulfilled(action.payload)),
+        catchError(error =>
+          of({
+            type: deleteUserRejected.type,
+            payload: error.xhr.response
+          })
+        )
+      )
+    )
+  );
+
+export { fetchUsersEpic, deleteUserEpic };
