+++ client/src/models/users/index.js
@@ -1,5 +1,5 @@
 import * as actions from "./actions";
-import epics from "./epics";
+import * as epics from "./epics";
 import { usersReducer as reducer } from "./reducers";
 
 export { actions, epics, reducer };
