+++ client/src/models/users/props.js
@@ -2,10 +2,11 @@ import _ from "lodash";
 
 //state
 const users = (state, payload) => ({ ...state, ..._.mapKeys(payload, "_id") });
+const afterUserDeleted = (state, payload) => _.omit(state, payload);
 //const error = errorObject => ({});
 
 //state selectors
 const usersSelector = ({ usersReducer }) =>
   usersReducer && Object.values(usersReducer);
 
-export { users, usersSelector };
+export { users, usersSelector, afterUserDeleted };
